+++ src/org/labkey/test/util/TestUser.java
@@ -47,9 +47,9 @@ public class TestUser
         getApiUserHelper().deleteUsers(false, _email);
     }
 
-    public Long getUserId()
+    public Integer getUserId()
     {
-        return (Long) getCreateUserResponse().getUserId();
+        return getCreateUserResponse().getUserId();
     }
 
     public String getEmail()
